+++ spec/support/typhoeus_macros.rb
@@ -1,0 +1,12 @@
+module TyphoeusMacros
+  def without_typhoeus_callbacks
+    before(:all) do
+      @original_typhoeus_callbacks = ::Typhoeus::Hydra.global_hooks.dup
+      ::Typhoeus::Hydra.clear_global_hooks
+    end
+
+    after(:all) do
+      ::Typhoeus::Hydra.global_hooks = @original_typhoeus_callbacks
+    end
+  end
+end