+++ spec/vcr_spec.rb
@@ -138,7 +138,8 @@ describe VCR do
 
     {
       :fakeweb  => VCR::HttpStubbingAdapters::FakeWeb,
-      :webmock  => VCR::HttpStubbingAdapters::WebMock
+      :webmock  => VCR::HttpStubbingAdapters::WebMock,
+      :faraday  => VCR::HttpStubbingAdapters::Faraday
     }.each do |symbol, klass|
       it "returns #{klass} for :#{symbol}" do
         VCR::Config.stub_with symbol